This commit was generated by cvs2svn to compensate for changes in r13821, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@13822 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/unix/x11/renderxv.c
@@ -61,7 +61,7 @@ int find_yuv_port(Display* display, XvPortID* port, fourcc_t* format)
   unsigned int version, release, request_base, event_base, error_base;
 
   /* XvQueryAdaptors */
-  int num_adaptors;
+  unsigned int num_adaptors;
   XvAdaptorInfo* adaptor_info = NULL;
   XvPortID port_id;
 
@@ -102,7 +102,7 @@ int find_yuv_port(Display* display, XvPortID* port, fourcc_t* format)
   }
 
   /* Find YUV capable adaptor. */
-  for (i = 0; i < num_adaptors; i++) {
+  for (i = 0; i < (int)num_adaptors; i++) {
     if (!(adaptor_info[i].type & XvInputMask
 	  && adaptor_info[i].type & XvImageMask))
     {
